Move handling of api key for detect_face_beta into detect-faces.js
