+++ modules/detect-faces.js
@@ -1,8 +1,16 @@
 var request = require('request');
+var vcapServices = require('vcap_services');
 
 module.exports.detectFaces = function(params, callback) {
+  var api_key = null
+  if(process.env.VCAP_SERVICES) {
+    var credentials = vcapServices.getCredentials('watson_vision_combined', null, 'visual-recognition-service')
+    api_key = credentials.api_key
+  } else {
+    api_key = process.env.VISUAL_RECOGNITION_API_KEY
+  }
   var options = {
-    url: `https://gateway-a.watsonplatform.net/visual-recognition/api/v3/detect_faces_beta?api_key=${params.api_key}&version=2016-05-17`,
+    url: `https://gateway-a.watsonplatform.net/visual-recognition/api/v3/detect_faces_beta?api_key=${api_key}&version=2016-05-17`,
   }
 
   if(params.images_file) {
